% Generated by roxygen2 (4.0.2): do not edit by hand
\name{barplot3d}
\alias{barplot3d}
\title{barplot3d}
\usage{
barplot3d(heights_arr, barcode = "0000", filename = "_3Dbarplot",
  column_colors = "default", bg_color = "white", color_types = FALSE,
  num_colors = 20, theta = -60, phi = 15, scale = 1, width = 700,
  height = 900, transp = 1, col_lab = NULL, row_lab = NULL,
  z_lab = NULL, bar_width = 0.5, bar_distance = 0.8,
  output_format = "screen")
}
\arguments{
\item{heigths_arr}{}
}
\description{
Plotting 3D Bars
}
\note{
todo: testing scale; average growth curve plot; 96 well overview plots
}
\examples{
barplot3d()
}
\keyword{plate}
\keyword{readers}

