% Generated by roxygen2 (4.0.2): do not edit by hand
\name{barPlotMP}
\alias{barPlotMP}
\title{Plotting kinetic data with Multiple Parameters with 2D Barplots}
\usage{
barPlotMP(kin_abs_df, slopes_df, fl_df, growth_df, barcode = "0000",
  wavelength = 0, lineColor = "red", overview = TRUE,
  singlePlot = FALSE, description = TRUE, markBest = FALSE,
  plotWidth = 7600, plotHeight = 4400, plotQuality = 70,
  filename = "_barPlotMP", plotRef = FALSE, output_format = "jpeg")
}
\arguments{
\item{filename}{}
}
\description{
Slopes need to be pre-calculated to also plot max and min slopes
}
\details{
barPlotMP
}
\note{
todo:  circles for high fl or high slope, still very ugly
}
\examples{
system.time(kinPlot(kin_df, lin_mod_df, wavelength=245, markBest=TRUE))
   good parameters f. jpg: plotWidth=7600, plotHeight=4400, plotQuality=70
   single plot: plotWidth=1200, plotHeight=800, plotQuality=70
}
\keyword{plate}
\keyword{readers}

