% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcAllLinModels}
\alias{calcAllLinModels}
\title{Calculate All Linear Models of given Data Frame}
\usage{
calcAllLinModels(kin_data_df, wavelength = 0, barcode = "0000")
}
\arguments{
\item{kin_data_df,}{wavelength}
}
\value{
data frame with
}
\description{
Calculates all possible linar models and stores coefficients in a data frame
}
\details{
calcAllLinModels
}
\note{
relativly well runtime optimized version - don't be confused about the ugly code: it's quite fast

todo: add barcode handling !!
}
\examples{
lin_mod_df <- calcAllLinModels(kin_df, wavelength=245)
}
\keyword{plate}
\keyword{readers}

