% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cubeFactory}
\alias{cubeFactory}
\title{Plotting single cubes}
\usage{
cubeFactory(cube_vec, cube_width, column_colors, num_colors = 20)
}
\arguments{
\item{cube_vec,}{cube_width, column_colors, num_colors=20}
}
\description{
cubeFactory
}
\note{
todo: nicer and faster
}
\examples{
importReaderData()
}
\keyword{for}
\keyword{microtiter}
\keyword{plates}
\keyword{plots}

