% Generated by roxygen2 (4.0.2): do not edit by hand
\name{growthExpressionPlot}
\alias{growthExpressionPlot}
\title{Plotting bacterial growth and Expression Data}
\usage{
growthExpressionPlot(growth_df, expr_df, wavelength = 660,
  growthColor = "darkgreen", exprColor = "red",
  plotTitle = "Av. Growth and Expression of Plate ", plotWidth = 1200,
  plotHeight = 1024, plotQuality = 70, filename = "growthExprPlot",
  output_format = "png")
}
\arguments{
\item{filename}{}
}
\description{
Slopes need to be pre-calculated to also plot max and min slopes
}
\details{
growthExpressionPlot
}
\note{
todo: connection line growth expression
}
\examples{
good parameters f. jpg: plotWidth=7600, plotHeight=4400, plotQuality=70
   single plot: plotWidth=1200, plotHeight=800, plotQuality=70
}
\keyword{plate}
\keyword{readers}

