% Generated by roxygen2 (4.0.2): do not edit by hand
\name{growthPlot}
\alias{growthPlot}
\title{Plotting Bacterial Growth Data}
\usage{
growthPlot(growth_df, wavelength = 660, lineColor = "red",
  plotTitle = "Average Growth of Plate ", plotWidth = 1200,
  plotHeight = 1024, plotQuality = 70, filename = "growthPlot",
  output_format = "png")
}
\arguments{
\item{filename}{}
}
\description{
Slopes need to be pre-calculated to also plot max and min slopes
}
\details{
growthPlot
}
\note{
todo: >=384 well format plotting,  WT handling, new text positioning
}
\examples{
good parameters f. jpg: plotWidth=7600, plotHeight=4400, plotQuality=70
   single plot: plotWidth=1200, plotHeight=800, plotQuality=70
}
\keyword{plate}
\keyword{readers}

