% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kinPlot}
\alias{kinPlot}
\title{Plotting Kinetic Absorption Data in Single files or Overview}
\usage{
kinPlot(kin_abs_df, slopes_df, barcode = "0000", wavelength = 0,
  lineColor = "red", overview = TRUE, singlePlot = FALSE,
  description = TRUE, markBest = FALSE, plotWidth = 7600,
  plotHeight = 4400, plotQuality = 70, filename = "_kinPlot",
  plotRef = FALSE, output_format = "png")
}
\arguments{
\item{kin_abs_df}{(data frame) - all absorption data}

\item{barcode}{(string)}

\item{wavelength}{(int)}
}
\description{
Slopes need to be pre-calculated to also plot max and min slopes
}
\details{
kinPlot
}
\note{
todo: >=384 well format plotting,  WT handling, new text positioning
}
\examples{
system.time(kinPlot(kin_df, lin_mod_df, wavelength=245, markBest=TRUE))
  system.time(kinPlot(kin_df, lin_mod_df, wavelength=245, markBest=TRUE, singlePlot=T, overview=F))
   good parameters f. jpg: plotWidth=7600, plotHeight=4400, plotQuality=70
   single plot: plotWidth=1200, plotHeight=800, plotQuality=70
}
\keyword{plate}
\keyword{readers}

