% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kinPlotMP}
\alias{kinPlotMP}
\title{Plotting Kinetic Data with Multiple Parameters (growth, activity, fluorescense, ratio act/fl)}
\usage{
kinPlotMP(kin_abs_df, slopes_df, fl_df, barcode = "0000", threshold = 1,
  wavelength = 0, overview = TRUE, singlePlot = FALSE,
  description = TRUE, markBest = FALSE, plotWidth = 7600,
  plotHeight = 4400, plotQuality = 70, filename = "_kinPlotMP",
  plotRef = FALSE, lineColor = "red", output_format = "png")
}
\arguments{
\item{kin_abs_df}{(data frame)}

\item{slopes_df}{(data frame)}

\item{fl_df}{(data frame)}
}
\description{
Slopes need to be pre-calculated to also plot max and min slopes
}
\details{
kinPlotMP
}
\note{
todo:  circles for high fl or high slope
}
\examples{
system.time(kinPlotMP(kin_df, lin_mod_df, wavelength=245, markBest=TRUE))
   good parameters f. jpg: plotWidth=7600, plotHeight=4400, plotQuality=70
   single plot: plotWidth=1200, plotHeight=800, plotQuality=70
}
\keyword{plate}
\keyword{readers}

