% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{platePlot}
\alias{platePlot}
\alias{platePlot-package}
\title{A library for 2D and 3D plotting of plate data reader files}
\description{
A library for 2D and 3D plotting of plate data reader files
}
\details{
A library for 2D and 3D plotting of plate data reader files
}
\note{
date: 141127

last modification: 141129
}
\author{
mark doerr \email{mark.doerr@uni-greifswald.de},
  and \email{mark@ismeralda.org}
}

