% Generated by roxygen2 (4.0.2): do not edit by hand
\name{platePlot3d}
\alias{platePlot3d}
\title{Plotting 3D bars of Measurements in Microtiter Format}
\usage{
platePlot3d(heights_df, barcode = "0000", wavelength = 0,
  use_slope = FALSE, filename = "_3Dbarplot", column_colors = "default",
  bg_color = "white", color_types = FALSE, num_colors = 20, theta = -60,
  phi = 15, scale = 1, width = 700, height = 900, transp = 1,
  col_lab = NULL, row_lab = NULL, z_lab = NULL, bar_width = 0.5,
  bar_distance = 0.8, output_format = "screen")
}
\arguments{
\item{heigths_arr}{}
}
\description{
Wrapper function for 3D bar plotting
}
\details{
platePlot3d
}
\note{
todo:
}
\examples{
platePlot3d()
}
\keyword{plate}
\keyword{readers}

