% Generated by roxygen2 (4.0.2): do not edit by hand
\name{selectBestLinModel}
\alias{selectBestLinModel}
\title{Select Best Linear Model for a given Data Frame}
\usage{
selectBestLinModel(kin_df)
}
\arguments{
\item{kin_data_df,}{wavelength}
}
\value{
data frame with
}
\description{
Calculates all possible linar models and stores coefficients in a data frame
     based on an idea of Martin Weiss
}
\details{
selectBestLinModel
}
\note{
relativly well runtime optimized version - don't be confused about the ugly code: it's quite fast

todo: not yet adapted to new data frame structure !!
}
\examples{
please wait until fixed !!!
}
\keyword{plate}
\keyword{readers}

